refactor: useMaxLoan to return amount or value
Prioritize liquidating col ratio error when removing collateral
